upgrading groovy version and converted to coldbox-module
