+++ Application.cfc
@@ -1,52 +1,50 @@
-﻿<!-----------------------------------------------------------------------
+﻿/**
 ********************************************************************************
 Copyright 2005-2007 ColdBox Framework by Luis Majano and Ortus Solutions, Corp
-www.coldbox.org | www.luismajano.com | www.ortussolutions.com
+www.coldboxframework.com | www.luismajano.com | www.ortussolutions.com
 ********************************************************************************
+*/
+component{
+	// Application properties
+	this.name = hash( getCurrentTemplatePath() );
+	this.sessionManagement = true;
+	this.sessionTimeout = createTimeSpan(0,0,30,0);
+	this.setClientCookies = true;
 
-Author     :	Luis Majano
-Date        :	10/16/2007
-Description :
-	This is the bootstrapper Application.cfc for ColdBox Applications.
-	It uses inheritance on the CFC, so if you do not want inheritance
-	then use the Application_noinheritance.cfc instead.
-	
-	
------------------------------------------------------------------------>
-<cfcomponent extends="coldbox.system.Coldbox" output="false">
-	<cfsetting enablecfoutputonly="yes">
-	<!--- APPLICATION CFC PROPERTIES --->
-	<cfset this.name = hash(getCurrentTemplatePath())> 
-	<cfset this.sessionManagement = true>
-	<cfset this.sessionTimeout = createTimeSpan(0,0,30,0)>
-	<cfset this.setClientCookies = true>
-	
-	<!--- COLDBOX STATIC PROPERTY, DO NOT CHANGE UNLESS THIS IS NOT THE ROOT OF YOUR COLDBOX APP --->
-	<cfset COLDBOX_APP_ROOT_PATH = getDirectoryFromPath(getCurrentTemplatePath())>
-	<!--- The web server mapping to this application. Used for remote purposes or static purposes --->
-	<cfset COLDBOX_APP_MAPPING   = "">
-	<!--- COLDBOX PROPERTIES --->
-	<cfset COLDBOX_CONFIG_FILE   = "">	
-	<!--- COLDBOX APPLICATION KEY OVERRIDE --->
-	<cfset COLDBOX_APP_KEY       = "">
-	<cfset this.mappings["/coldbox"] 	= COLDBOX_APP_ROOT_PATH & "coldbox">
-	
-	<!--- on Request Start --->
-	<cffunction name="onRequestStart" returnType="boolean" output="true">
-		<!--- ************************************************************* --->
-		<cfargument name="targetPage" type="string" required="true" />
-		<!--- ************************************************************* --->
-		
-		<!--- Process A ColdBox Request Only --->
-		<cfif findNoCase('index.cfm', listLast(arguments.targetPage, '/'))>
-			<!--- Reload Checks --->
-			<cfset reloadChecks()>
-			<!--- Process Request --->
-			<cfset processColdBoxRequest()>
-		</cfif>
-			
-		<!--- WHATEVER YOU WANT BELOW --->
-		<cfreturn true>
-	</cffunction>
-	
-</cfcomponent>
+	// COLDBOX STATIC PROPERTY, DO NOT CHANGE UNLESS THIS IS NOT THE ROOT OF YOUR COLDBOX APP
+	COLDBOX_APP_ROOT_PATH = getDirectoryFromPath( getCurrentTemplatePath() );
+	// The web server mapping to this application. Used for remote purposes or static purposes
+	COLDBOX_APP_MAPPING   = "";
+	// COLDBOX PROPERTIES
+	COLDBOX_CONFIG_FILE 	 = "";
+	// COLDBOX APPLICATION KEY OVERRIDE
+	COLDBOX_APP_KEY 		 = "";
+
+	// application start
+	public boolean function onApplicationStart(){
+		application.cbBootstrap = new coldbox.system.Bootstrap( COLDBOX_CONFIG_FILE, COLDBOX_APP_ROOT_PATH, COLDBOX_APP_KEY, COLDBOX_APP_MAPPING );
+		application.cbBootstrap.loadColdbox();
+		return true;
+	}
+
+	// request start
+	public boolean function onRequestStart(String targetPage){
+		// Process ColdBox Request
+		application.cbBootstrap.onRequestStart( arguments.targetPage );
+
+		return true;
+	}
+
+	public void function onSessionStart(){
+		application.cbBootStrap.onSessionStart();
+	}
+
+	public void function onSessionEnd( struct sessionScope, struct appScope ){
+		arguments.appScope.cbBootStrap.onSessionEnd( argumentCollection=arguments );
+	}
+
+	public boolean function onMissingTemplate( template ){
+		return application.cbBootstrap.onMissingTemplate( argumentCollection=arguments );
+	}
+
+}
