+++ config/Application.cfm
@@ -1,4 +1,4 @@
-﻿<!-----------------------------------------------------------------------
+<!-----------------------------------------------------------------------
 Template :  Application.cfm 
 Author 	 :	Luis Majano
 Date     :	January 22, 2006
