+++ config/Coldbox.cfc
@@ -1,70 +1,29 @@
-﻿<cfcomponent output="false" hint="My App Configuration">
+<cfcomponent output="false" hint="My App Configuration">
 <cfscript>
-/**
-structures/arrays to create for configuration
-
-- coldbox (struct)
-- settings (struct)
-- conventions (struct)
-- environments (struct)
-- wirebox (struct)
-- ioc (struct)
-- debugger (struct)
-- mailSettings (struct)
-- i18n (struct)
-- webservices (struct)
-- datasources (struct)
-- layoutSettings (struct)
-- layouts (array of structs)
-- cacheBox (struct)
-- interceptorSettings (struct)
-- interceptors (array of structs)
-- modules (struct)
-- logBox (struct)
-- flash (struct)
-- orm (struct)
-- validation (struct)
-
-Available objects in variable scope
-- controller
-- logBoxConfig
-- appMapping (auto calculated by ColdBox)
-
-Required Methods
-- configure() : The method ColdBox calls to configure the application.
-Optional Methods
-- detectEnvironment() : If declared the framework will call it and it must return the name of the environment you are on.
-- {environment}() : The name of the environment found and called by the framework.
-
-*/
-
 	// Configure ColdBox Application
 	function configure(){
 
 		// coldbox directives
 		coldbox = {
 			//Application Setup
-			appName 				= "Your app name here",
-			eventName 				= "event",
+			appName 				= "Development Shell",
 
 			//Development Settings
-			debugMode				= true,
-			debugPassword			= "",
 			reinitPassword			= "",
 			handlersIndexAutoReload = true,
 
 			//Implicit Events
 			defaultEvent			= "General.index",
-			requestStartHandler		= "Main.onRequestStart",
+			requestStartHandler		= "",
 			requestEndHandler		= "",
-			applicationStartHandler = "Main.onAppInit",
+			applicationStartHandler = "main.onAppInit",
 			applicationEndHandler	= "",
 			sessionStartHandler 	= "",
 			sessionEndHandler		= "",
 			missingTemplateHandler	= "",
 
 			//Extension Points
-			UDFLibraryFile 				= "includes/helpers/ApplicationHelper.cfm",
+			ApplicationHelper 			= "",
 			coldboxExtensionsLocation 	= "",
 			modulesExternalLocation		= [],
 			pluginsExternalLocation 	= "",
@@ -76,7 +35,7 @@ Optional Methods
 			//Error/Exception Handling
 			exceptionHandler		= "",
 			onInvalidEvent			= "",
-			customErrorTemplate		= "",
+			customErrorTemplate		= "/coldbox/system/includes/BugReport.cfm",
 
 			//Application Aspects
 			handlerCaching 			= false,
@@ -86,34 +45,32 @@ Optional Methods
 
 		// custom settings
 		settings = {
-
 		};
 
-		// environment settings, create a detectEnvironment() method to detect it yourself.
-		// create a function with the name of the environment so it can be executed if that environment is detected
-		// the value of the environment is a list of regex patterns to match the cgi.http_host.
-		environments = {
-			//development = "^cf8.,^railo."
-		};
+		//goovy files paths
+		groovy = {libPath='/models/groovy,/models/anotherPath'};
+
+		// Activate WireBox
+		wirebox = { enabled = true, singletonReload=false };
 
 		// Module Directives
 		modules = {
-			//Turn to false in production
-			autoReload = false,
-			// An array of modules names to load, empty means all of them
-			include = [],
-			// An array of modules names to NOT load, empty means none
-			exclude = []
+			//Turn to false in production, on for dev
+			autoReload = false
 		};
 
 		//LogBox DSL
 		logBox = {
 			// Define Appenders
 			appenders = {
-				coldboxTracer = { class="coldbox.system.logging.appenders.ColdboxTracerAppender" }
+				files={class="coldbox.system.logging.appenders.RollingFileAppender",
+					properties = {
+						filename = "groovyloader", filePath="/#appMapping#/logs"
+					}
+				}
 			},
 			// Root Logger
-			root = { levelmax="INFO", appenders="*" },
+			root = { levelmax="DEBUG", appenders="*" },
 			// Implicit Level Categories
 			info = [ "coldbox.system" ]
 		};
@@ -123,143 +80,14 @@ Optional Methods
 			defaultLayout = "Main.cfm"
 		};
 
-		//Interceptor Settings
-		interceptorSettings = {
-			throwOnInvalidStates = false,
-			customInterceptionPoints = ""
-		};
-
 		//Register interceptors as an array, we need order
 		interceptors = [
 			//SES
 			{class="coldbox.system.interceptors.SES",
 			 properties={}
-			},
-			//Groovy Loader
-			//Paths that hold groovy libs
-			//Paths that hold jar's for us to load automagically
-			{class="plugins.GroovyLoader.GroovyStarter",
-			 properties={groovyLibPaths='model/groovy,model/anotherPath', javalibPaths='model/lib'}
 			}
 		];
-
-		// Object & Form Validation
-		validation = {
-			// manager = "class path" or none at all to use ColdBox as the validation manager
-			// The shared constraints for your application.
-			sharedConstraints = {
-				// EX
-				// myForm = { name={required=true}, age={type="numeric",min="18"} }
-			}
-		};
-
-
-		/*
-
-		// ORM services, injection, etc
-		orm = {
-			// entity injection
-			injection = {
-				// enable it
-				enabled = true,
-				// the include list for injection
-				include = "",
-				// the exclude list for injection
-				exclude = ""
-			}
-		};
-
-		// flash scope configuration
-		flash = {
-			scope = "session,client,cluster,ColdboxCache,or full path",
-			properties = {}, // constructor properties for the flash scope implementation
-			inflateToRC = true, // automatically inflate flash data into the RC scope
-			inflateToPRC = false, // automatically inflate flash data into the PRC scope
-			autoPurge = true, // automatically purge flash data for you
-			autoSave = true // automatically save flash scopes at end of a request and on relocations.
-		};
-
-		//Register Layouts
-		layouts = [
-			{ name = "login",
-		 	  file = "Layout.tester.cfm",
-			  views = "vwLogin,test",
-			  folders = "tags,pdf/single"
-			}
-		];
-
-		//Conventions
-		conventions = {
-			handlersLocation = "handlers",
-			pluginsLocation  = "plugins",
-			viewsLocation 	 = "views",
-			layoutsLocation  = "layouts",
-			modelsLocation 	 = "model",
-			eventAction 	 = "index"
-		};
-
-		//IOC Integration
-		ioc = {
-			framework 		= "lightwire",
-			reload 	  	  	= true,
-			objectCaching 	= false,
-			definitionFile  = "config/coldspring.xml.cfm",
-			parentFactory 	= {
-				framework = "coldspring",
-				definitionFile = "config/parent.xml.cfm"
-			}
-		};
-
-		//Debugger Settings
-		debugger = {
-			enableDumpVar = false,
-			persistentRequestProfilers = true,
-			maxPersistentRequestProfilers = 10,
-			maxRCPanelQueryRows = 50,
-			showRCSnapshots = false,
-			//Panels
-			showTracerPanel = true,
-			expandedTracerPanel = true,
-			showInfoPanel = true,
-			expandedInfoPanel = true,
-			showCachePanel = true,
-			expandedCachePanel = true,
-			showRCPanel = true,
-			expandedRCPanel = true,
-			showModulesPanel = true,
-			expandedModulesPanel = false
-		};
-
-		//Mailsettings
-		mailSettings = {
-			server = "",
-			username = "",
-			password = "",
-			port = 25
-		};
-
-		//i18n & Localization
-		i18n = {
-			defaultResourceBundle = "includes/i18n/main",
-			defaultLocale = "en_US",
-			localeStorage = "session",
-			unknownTranslation = "**NOT FOUND**"
-		};
-
-		//webservices
-		webservices = {
-			testWS = "http://www.test.com/test.cfc?wsdl",
-			AnotherTestWS = "http://www.coldbox.org/distribution/updatews.cfc?wsdl"
-		};
-
-		//Datasources
-		datasources = {
-			mysite   = {name="mySite", dbType="mysql", username="root", password="pass"},
-			blog_dsn = {name="myBlog", dbType="oracle", username="root", password="pass"}
-		};
-		*/
 
 	}
-
 </cfscript>
 </cfcomponent>
