--- model/groovy/Test.groovy
@@ -1,4 +1,0 @@
-//incoming varCollection
-def myArray = [1,2,3,4]
-//add to varCollection
-varCollection['MyArray'] =  myArray
