--- model/groovy/util/Phone.groovy
@@ -1,6 +1,0 @@
-package util
-
-class Phone{
-	def phone = "";
-	def areacode = "";
-}
