+++ models/groovy/Test.groovy
@@ -1,0 +1,4 @@
+//incoming varCollection
+def myArray = [1,2,3,4]
+//add to varCollection
+varCollection['MyArray'] =  myArray
