+++ models/model/groovy/util/Phone.groovy
@@ -1,0 +1,6 @@
+package util
+
+class Phone{
+	def phone = "";
+	def areacode = "";
+}
