+++ modules/cbgroovy/handlers/Home.cfc
@@ -1,0 +1,10 @@
+/**
+* A normal ColdBox Event Handler
+*/
+component{
+
+	function index(event,rc,prc){
+		event.setView( "home/index" );
+	}
+
+}
