+++ modules/cbgroovy/views/home/index.cfm
@@ -1,0 +1,1 @@
+<h1>Welcome to my cool module page!</h1>
